backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
chromadb==0.4.18
ollama==0.1.6
numpy==1.24.3
pydub==0.25.1
moviepy==1.0.3
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1